Fixed crash when editing effects without a gif Fixed LUB colors using vec4 instead of vec3

+++ browedit/BrowEdit.cpp
@@ -314,7 +314,7 @@ void BrowEdit::run()
 			if (editMode == EditMode::Texture && activeMapView)
 			{
 				if (ImGui::IsKeyPressed(GLFW_KEY_LEFT_BRACKET))
-					activeMapView->textureSelected = (activeMapView->textureSelected + activeMapView->map->rootNode->getComponent<Gnd>()->textures.size() - 1) % activeMapView->map->rootNode->getComponent<Gnd>()->textures.size();
+					activeMapView->textureSelected = (activeMapView->textureSelected + activeMapView->map->rootNode->getComponent<Gnd>()->textures.size() - 1) % (int)activeMapView->map->rootNode->getComponent<Gnd>()->textures.size();
 				if (ImGui::IsKeyPressed(GLFW_KEY_RIGHT_BRACKET))
 					activeMapView->textureSelected = (activeMapView->textureSelected + 1) % activeMapView->map->rootNode->getComponent<Gnd>()->textures.size();
